update logging demo
