+++ README.md
@@ -29,7 +29,7 @@ This repository contains examples for common use cases and API implementations. 
 | [use-interop](./how-to/use-interop/)                                 | This example contains various implementations and use cases for the Interop API. Including Interop Broker to Interop Broker communication on the same machine and cross machine/platform communication using our Cloud Interop solution. | [Example](https://start.openfin.co/?manifest=https%3A%2F%2Fbuilt-on-openfin.github.io%2Fcontainer-starter%2Fv38%2Fuse-interop-setup-multi-platform-interop%2Fmanifest.fin.json)                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       |
 | [use-a-manifest](./how-to/use-a-manifest/)                           | This example contains various examples of how to use the OpenFin manifest                                                                                                                                                                | [Example](https://start.openfin.co/?manifest=https%3A%2F%2Fbuilt-on-openfin.github.io%2Fcontainer-starter%2Fv38%2Fuse-a-manifest-create-a-single-page-platform%2Fmanifest.fin.json)                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   |
 | [use-content-creation-rules](./how-to/use-content-creation-rules/)   | This example demonstrates how content creation rules can be applied                                                                                                                                                                      | [Example](https://start.openfin.co/?manifest=https%3A%2F%2Fbuilt-on-openfin.github.io%2Fcontainer-starter%2Fv38%2Fuse-content-creation-rules%2Fmanifest.fin.json)                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     |
-| [use-logging-apis](./how-to/use-logging-apis/)                       | This example shows how to use OpenFin APIs to retrieve logs and send them to a server.                                                                                                                                                   |                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       |
+| [use-logging](./how-to/use-logging)                       | This example shows how to use OpenFin APIs to retrieve logs and send them to a server.                                                                                                                                                   |                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       |
 | [use-launch-external-process](./how-to/use-launch-external-process/) | This example demonstrates how to use the launch external process secure api.                                                                                                                                                             | [Example](https://start.openfin.co/?manifest=https%3A%2F%2Fbuilt-on-openfin.github.io%2Fcontainer-starter%2Fv38%2Fuse-launch-external-process%2Fmanifest.fin.json)                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    |
 | [use-licensed-codecs](./how-to/use-licensed-codecs/)                 | This example demonstrates how to use the new licensed codec support available in v38+ versions of the runtime.                                                                                                                           |                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       |
