--- how-to/use-logging-apis/client/tsconfig.json
@@ -1,15 +1,0 @@
-{
-	"compilerOptions": {
-		"target": "es2021",
-		"module": "ES2020",
-		"sourceMap": true,
-		"rootDir": "./src",
-		"outDir": "build",
-		"skipLibCheck": true,
-		"strict": true,
-		"esModuleInterop": true,
-		"forceConsistentCasingInFileNames": true,
-		"moduleResolution": "node"
-	},
-	"include": ["./src/**/*.ts"]
-}
