--- how-to/use-logging-apis/client/types/fin.d.ts
@@ -1,5 +1,0 @@
-import type { fin as FinApi } from "@openfin/core";
-
-declare global {
-	const fin: typeof FinApi;
-}
