--- how-to/use-logging-apis/package.json
@@ -1,37 +1,0 @@
-{
-	"name": "use-logging-apis",
-	"version": "40.104.2",
-	"description": "A demonstration of how to use logging apis",
-	"main": "index.js",
-	"scripts": {
-		"kill": "node ./scripts/kill.mjs",
-		"client": "node ./scripts/launch.mjs",
-		"server": "node ./server/build/index.js",
-		"build-client": "webpack build --config ./client/webpack.config.js --mode=development",
-		"build-server": "tsc --project ./server",
-		"build": "npm run build-client && npm run build-server",
-		"start": "npm run server",
-		"setup": "cd ../../ && npm install && cd how-to/use-logging-apis && npm run build"
-	},
-	"author": "martyn.janes@openfin.co",
-	"license": "SEE LICENSE IN LICENSE.MD",
-	"devDependencies": {
-		"@openfin/core": "40.104.2",
-		"@types/multer": "^1.4.11",
-		"@types/node": "^20.12.11",
-		"body-parser": "^1.20.2",
-		"express": "^4.19.2",
-		"@openfin/node-adapter": "40.104.2",
-		"source-map-loader": "^5.0.0",
-		"ts-loader": "^9.5.1",
-		"typescript": "^5.4.5",
-		"webpack": "^5.91.0",
-		"webpack-cli": "^5.1.4"
-	},
-	"dependencies": {
-		"connect-busboy": "^1.0.0",
-		"multer": "^1.4.5-lts.1",
-		"@openfin/core": "40.104.2",
-		"@openfin/node-adapter": "40.104.2"
-	}
-}
