--- how-to/use-logging-apis/server/tsconfig.json
@@ -1,15 +1,0 @@
-{
-	"compilerOptions": {
-		"target": "es2021",
-		"module": "commonjs",
-		"sourceMap": true,
-		"rootDir": "./src",
-		"outDir": "build",
-		"skipLibCheck": true,
-		"strict": true,
-		"esModuleInterop": true,
-		"forceConsistentCasingInFileNames": true,
-		"moduleResolution": "node"
-	},
-	"include": ["./src/**/*.ts"]
-}
