+++ how-to/use-logging/use-log-uploader-with-self-hosted-service/.vscode/settings.json
@@ -1,0 +1,6 @@
+{
+    "editor.codeActionsOnSave": {
+        "source.fixAll.eslint": "always",
+        "source.fixAll": "always"
+      },
+}
