+++ how-to/use-logging/use-log-uploader-with-self-hosted-service/package.json
@@ -1,0 +1,28 @@
+{
+	"name": "log-uploader-api",
+	"version": "40.104.2",
+	"main": "server.js",
+	"scripts": {
+		"start": "npx tsx server.ts"
+	},
+	"keywords": [],
+	"author": "Ameet Jayawant",
+	"license": "SEE LICENSE IN LICENSE.MD",
+	"description": "",
+	"dependencies": {
+		"@types/multer": "^1.4.12",
+		"body-parser": "^1.20.3",
+		"cors": "^2.8.5",
+		"express": "^4.21.2",
+		"multer": "^1.4.5-lts.1"
+	},
+	"devDependencies": {
+		"nodemon": "^3.1.9",
+		"prettier-eslint": "^16.3.0",
+		"ts-loader": "^9.5.2",
+		"tsx": "^4.19.3",
+		"typescript": "^5.8.2",
+		"webpack": "^5.98.0",
+		"webpack-cli": "^6.0.1"
+	}
+}
