+++ how-to/use-logging/use-logging-apis/client/types/fin.d.ts
@@ -1,0 +1,5 @@
+import type { fin as FinApi } from "@openfin/core";
+
+declare global {
+	const fin: typeof FinApi;
+}
