+++ how-to/use-logging/use-logging-apis/public/manifest.fin.json
@@ -1,0 +1,64 @@
+{
+	"runtime": {
+		"arguments": "--enable-mesh --security-realm=use-application-logging",
+		"version": "40.130.104.2"
+	},
+	"licenseKey": "openfin-demo-license-key",
+	"platform": {
+		"name": "use-application-logging",
+		"uuid": "use-application-logging",
+		"description": "An example of using application logging",
+		"providerUrl": "http://localhost:5050/html/provider.html",
+		"icon": "http://localhost:5050/favicon.ico",
+		"autoShow": true,
+		"defaultWidth": 1200,
+		"defaultHeight": 800,
+		"defaultCentered": true,
+		"minWidth": 800,
+		"minHeight": 400,
+		"appLogsTimezone": "utc",
+		"enableAppLogging": true,
+		"logManagement": {
+			"enabled": true,
+			"url": "http://localhost:5050"
+		}
+	},
+	"snapshot": {
+		"windows": [
+			{
+				"defaultHeight": 900,
+				"defaultWidth": 1200,
+				"defaultCentered": true,
+				"layout": {
+					"content": [
+						{
+							"type": "row",
+							"content": [
+								{
+									"type": "component",
+									"componentName": "view",
+									"componentState": {
+										"url": "http://localhost:5050/html/app.html"
+									}
+								},
+								{
+									"type": "component",
+									"componentName": "view",
+									"componentState": {
+										"url": "http://localhost:5050/html/app.html"
+									}
+								}
+							]
+						}
+					]
+				}
+			}
+		]
+	},
+	"supportInformation": {
+		"company": "OpenFin",
+		"product": "Container Logging APIs Example",
+		"email": "support@openfin.co",
+		"forwardErrorReports": false
+	}
+}
