+++ how-to/use-logging/use-logging-apis/public/uploads/applogs/AppLogs.md
@@ -1,0 +1,4 @@
+# App Logs
+
+This directory will be written to by our example node server. It will receive the app log files that are sent to the server by the RVM.
+The file will be a zip and it will contain the app.log file which has the combined data from all web pages writing to the console.
