+++ how-to/use-logging/use-logging-apis/public/uploads/runtimelogs/SystemLogs.md
@@ -1,0 +1,3 @@
+# System Logs
+
+This directory will be written to by our example node server. It will receive the debug log file that contains runtime messages as well as other data and it will write it to this directory.
